+++ routes/projects.js
@@ -250,6 +250,12 @@ router.delete('/project/:id', async function (req, res) {
         return res.status(401).send({ message: 'Not authorized' });
     }
 
+    let users_in_project = await dbsrv.mongo_users().find({ projects: req.params.id }).toArray();
+
+    if (users_in_project.length > 0){
+      return res.status(403).send({ message: 'Project is not empty. Cannot remove' });
+    }
+
     try {
         await prjsrv.remove_project(req.params.id, user.uid);
     } catch (e) {