+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Ensure emails have name, destination and subject
 * Fix bug in TP account deletion
+* Fix database creation validation confirmation email
 
 ## 1.4.32 (2025-04-22)