+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Ensure emails have name, destination and subject
 * Fix bug in TP account deletion
+* Adds a 'last_extended' attribute to the Project type
 
 ## 1.4.32 (2025-04-22)