+++ CHANGELOG.md
@@ -12,7 +12,9 @@
 * Remove DB 'size' attribute
 * Allow admins to add/remove project managers
 * Fix add/remove manager http returns
+* Modified add_readme template
 * Fix typo in cron due to remove_tp_reservation changes
+* Allow spaces, - and ' in user first/last names
 
 ## 1.4.32 (2025-04-22)