+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Modified add_readme template
 * Fix typo in cron due to remove_tp_reservation changes
 * Allow spaces, - and ' in user first/last names
+* Fix project deletion process
 
 ## 1.4.32 (2025-04-22)